Work on stream serializers

+++ src/lib/mint/garbagecollector.cpp
@@ -5,5 +5,5 @@ using namespace mint;
 
 MINT_FUNCTION(mint_garbage_collector_collect, 0, cursor) {
 	FunctionHelper helper(cursor, 0);
-	helper.returnValue(create_number(GarbageCollector::instance().collect()));
+	helper.returnValue(create_number(static_cast<double>(GarbageCollector::instance().collect())));
 }
